Fixes sync_all.sh path error.
When developing the sync_all.sh script, we always ran it from the
directory where it resides. Unfortunately, that made us miss the fact
that the call to './fork_sync.sh' was incorrect. It should have been
something along the lines of 'path_from_caller_pwd/fork_sync.sh'.

This commit fixes that. It has been verified by running the script
from a different directory. :)

Change-Id: Ic65f7172aaf535e6b36dc9ce545b82785b152f8a