Adds a small manager documentation generator script.
In order to more easily keep the manager docs up to date, we have
started moving them from Open Office / Apple Pages to Markdown.

To fulfil the promise of keeping the docs up to date, this script uses
Ruby's dir-globbing features to recursively find and transform into
HTML all .markdown-files starting in the given directory.

The goal is to automatically generate the docs on the jenkins server
from the master branch every night.

Change-Id: I1028885430d7f23975918f94002e5ba786caa88a